+++ src/lib.rs
@@ -5,6 +5,7 @@ pub mod deserializers;
 pub mod embeddings;
 pub mod error;
 pub mod frameworks;
+pub mod gemini;
 pub mod indexes;
 pub mod kg_extractor;
 pub mod prompt_critiques;
@@ -13,6 +14,7 @@ pub mod prompts;
 pub mod schemas;
 pub mod serializers;
 pub mod server;
+pub mod sessions;
 pub mod tools;
 pub mod utils;